updates from main rig;
 Conflicts:
	.idea/workspace.xml

+++ db/sqlite/datamodels/sql/CREATETABLE/ProgramFinancingSchedule.sql
@@ -16,9 +16,9 @@ CREATE TABLE IF NOT EXISTS "ProgramFinancingSchedule"
 	"AccountSequence"	TEXT(80) NULL DEFAULT 'NS',
 	"AgencyName"	TEXT(80) NULL DEFAULT 'NS',
 	"FundName"	TEXT(80) NULL DEFAULT 'NS',
-	"OriginalAmount"	NUMERIC NULL DEFAULT 0,
-	"BudgetAmount"	NUMERIC NULL DEFAULT 0,
-	"AgencyAmount"	NUMERIC NULL DEFAULT 0,
-	"Amount"	NUMERIC NULL DEFAULT 0,
+	"OriginalAmount"	DOUBLE NULL DEFAULT 0.0,
+	"BudgetAmount"	DOUBLE NULL DEFAULT 0.0,
+	"AgencyAmount"	DOUBLE NULL DEFAULT 0.0,
+	"Amount"	DOUBLE NULL DEFAULT 0.0,
 	PRIMARY KEY("ProgramFinancingScheduleId" AUTOINCREMENT)
 );
